shadow_shim_helper_rs::explicit_drop
Struct ExplicitDropper

pub struct ExplicitDropper<DropFn, Value>
where
    DropFn: FnOnce(Value),
{ /* private fields */ }

Wrapper that uses a provided function to drop the inner value.

This is helpful for working with locals that must be explicitly dropped, without having to explicitly do so at every potential exit point of the current function.

fn validate_and_rc(root: &Root, s: String) -> Result<RootedRc<String>, Box<dyn Error>> {
  let rc_string = ExplicitDropper::new(
    RootedRc::new(root, s),
    |value| value.explicit_drop(root));

  if !rc_string.starts_with("x") {
    // `ExplicitDropper` will call the provided closure, safely dropping the RootedRc.
    return Err("bad prefix".into());
  }

  // We extract the value from the dropper at the point where we transfer ownership;
  // in this case the closure is never called.
  return Ok(rc_string.into_value())
}

Implementations

impl<DropFn, Value> ExplicitDropper<DropFn, Value>
where DropFn: FnOnce(Value),

pub fn new(value: Value, dropper: DropFn) -> Self

Create a wrapped value

pub fn into_value(self) -> Value

Unwrap the value, discarding the dropper.