shadow_shim_helper_rs::rootedcell::refcell

Struct RootedRefCell

#[repr(C)]
pub struct RootedRefCell<T> { /* private fields */ }

Analagous to core::cell::RefCell. In particular like core::cell::RefCell and unlike std::sync::Mutex, it doesn’t perform any atomic operations internally, making it relatively inexpensive.

Unlike core::cell::RefCell, this type is Send and Sync if T is Send. This is safe because the owner is required to prove access to the associated Root, which is ![Sync], to borrow.

Implementations

impl<T> RootedRefCell<T>

pub fn new(root: &Root, val: T) -> Self

Create a RootedRefCell associated with root.

pub fn borrow<'a>(&'a self, root: &'a Root) -> RootedRefCellRef<'a, T>

Borrow a reference. Panics if root is for the wrong Root, or if this object is alread mutably borrowed.

pub fn borrow_mut<'a>(&'a self, root: &'a Root) -> RootedRefCellRefMut<'a, T>

Borrow a mutable reference. Panics if root is for the wrong Root, or if this object is already borrowed.

pub fn into_inner(self) -> T