shadow_shim_helper_rs

Module simulation_time

Values for working with a simulated duration. Use EmulatedTime to represent an instant in time.

In Rust, use EmulatedTime to represent an instant in time, or SimulationTime to represent a time interval. SimulationTime is meant to replace SimulationTime from the C APIs.

This module contains some identically-named constants defined as C macros in main/core/definitions.h.

Modules

export

Structs

SimulationTime

Constants

SIMTIME_INVALID

Invalid simulation time.

SIMTIME_MAX

Maximum and minimum valid values.

SIMTIME_MIN

SIMTIME_ONE_HOUR

Represents one hour in simulation time.

SIMTIME_ONE_MICROSECOND

Represents one microsecond in simulation time.

SIMTIME_ONE_MILLISECOND

Represents one millisecond in simulation time.

SIMTIME_ONE_MINUTE

Represents one minute in simulation time.

SIMTIME_ONE_NANOSECOND

Represents one nanosecond in simulation time.

SIMTIME_ONE_SECOND

Represents one second in simulation time.

Type Aliases

CSimulationTime

The same as the type alias in the ‘main/cshadow.rs’ bindings.