shadow_shim_helper_rs
Module syscall_types

Structs

ForeignPtr

Represents a pointer to a virtual address in plugin memory. The pointer is not guaranteed to be valid or aligned.

ManagedPhysicalMemoryAddr

Represents a pointer to a physical address in plugin memory.

SyscallArgs

Type Aliases

UntypedForeignPtr

Used to indicate an untyped ForeignPtr in C code.

Unions

SyscallReg

A register used for input/output in a syscall.