shadow_shim_helper_rs::syscall_types
Type Alias UntypedForeignPtr

pub type UntypedForeignPtr = ForeignPtr<()>;

Used to indicate an untyped ForeignPtr in C code.

We use the unit type as the generic rather than libc::c_void since the unit type is zero-sized and libc::c_void has a size of 1 byte, and this prevents us from accidentally accessing an “untyped” pointer.

Aliased Type

struct UntypedForeignPtr { /* private fields */ }