shadow_shim_helper_rs

Module emulated_time

Deals with instances of time in a Shadow simulation.

Modules

export

Structs

AtomicEmulatedTime

EmulatedTime

An instant in time (analagous to core::time::Instant) in the Shadow simulation.

Constants

EMUTIME_INVALID

EMUTIME_MAX

EMUTIME_MIN

EMUTIME_SIMULATION_START

The number of nanoseconds from the epoch to January 1st, 2000 at 12:00am UTC. This is used to emulate to applications that we are in a recent time.

EMUTIME_UNIX_EPOCH

Duplicated as EmulatedTime::UNIX_EPOCH

SIMULATION_START_SEC

Type Aliases

CEmulatedTime

Emulation time in nanoseconds.