shadow_shim_helper_rs::emulated_time

Struct EmulatedTime

#[repr(C)]
pub struct EmulatedTime(/* private fields */);

An instant in time (analagous to core::time::Instant) in the Shadow simulation.

Implementations

impl EmulatedTime

pub const SIMULATION_START: Self

The start time of the simulation - 00:00:00 UTC on 1 January, 2000.

pub const UNIX_EPOCH: Self

The Unix epoch (00:00:00 UTC on 1 January 1970)

pub const MAX: Self

pub const MIN: Self

pub const fn from_c_emutime(val: CEmulatedTime) -> Option<Self>

Get the instance corresponding to val SimulationTime units since the Unix Epoch.

pub const fn to_c_emutime(val: Option<Self>) -> CEmulatedTime

Convert to number of SimulationTime units since the Unix Epoch.

pub fn from_abs_simtime(val: SimulationTime) -> Self

Get the instant corresponding to val time units since the simulation began.

pub fn to_abs_simtime(self) -> SimulationTime

Convert to the SimulationTime since the simulation began.

pub fn duration_since(&self, earlier: &EmulatedTime) -> SimulationTime

Returns the duration since earlier, or panics if earlier is after self, or if the difference can’t be represented as SimulationTime.

pub fn checked_duration_since( &self, earlier: &EmulatedTime, ) -> Option<SimulationTime>

Returns the duration since earlier, or None if earlier is after self.

pub fn saturating_duration_since( &self, earlier: &EmulatedTime, ) -> SimulationTime

Returns the duration since earlier, or 0 if earlier is after self.

pub fn checked_add(&self, duration: SimulationTime) -> Option<EmulatedTime>

pub fn checked_sub(&self, duration: SimulationTime) -> Option<EmulatedTime>

pub fn saturating_add(&self, duration: SimulationTime) -> EmulatedTime

pub fn saturating_sub(&self, duration: SimulationTime) -> EmulatedTime