Module shadow_shim_helper_rs::rootedcell

Modules

cell

rc

refcell

Structs

Root

A Root is a ![Sync] token. Proof of access to a Root is used to inexpensively ensure safety of safety in rc::RootedRc and refcell::RootedRefCell.

Tag

Every object root is assigned a Tag, which we ensure is globally unique. Each Tag value uniquely identifies a Root.