Module shadow_shim_helper_rs::shim_shmem::export

Functions

shimshmem_getBlockedSignals⚠

Safety

shimshmem_getEmulatedTime⚠

Safety

shimshmem_getLogLevel⚠

Safety

shimshmem_getLoggingStartTime⚠

Get the logging start time

shimshmem_getMaxRunaheadTime⚠

Safety

shimshmem_getModelUnblockedSyscallLatency⚠

Get whether to model latency of unblocked syscalls.

shimshmem_getProcessStraceFd⚠

Safety

shimshmem_getShadowPid⚠

Safety

shimshmem_getSigAltStack⚠

Get the signal stack as set by sigaltstack(2).

shimshmem_getSignalAction⚠

Safety

shimshmem_getThreadId⚠

Safety

shimshmem_getThreadPendingSignals⚠

Safety

shimshmem_getTscHz⚠

Safety

shimshmem_getUnappliedCpuLatency⚠

Safety

shimshmem_incrementUnappliedCpuLatency⚠

Safety

shimshmem_maxUnappliedCpuLatency⚠

Get the configured maximum unblocked syscall latency to accumulate before yielding.

shimshmem_resetUnappliedCpuLatency⚠

Safety

shimshmem_setBlockedSignals⚠

Set the process’s pending signal set.

shimshmem_setEmulatedTime⚠

Safety

shimshmem_setMaxRunaheadTime⚠

Safety

shimshmem_setSigAltStack⚠

Set the signal stack as set by sigaltstack(2).

shimshmem_setSignalAction⚠

Safety

shimshmem_setThreadPendingSignals⚠

Set the process’s pending signal set.

shimshmem_setThreadSiginfo⚠

Set the siginfo for the given signal number.

shimshmem_unblockedSyscallLatency⚠

Get the configured latency to emulate for each unblocked syscall.

shimshmem_unblockedVdsoLatency⚠

Get the configured latency to emulate for each unblocked vdso “syscall”.

shimshmemhost_lock⚠

Safety

shimshmemhost_unlock⚠

Safety

shimshmemthread_size

Type Aliases

ShimShmemHost

ShimShmemHostLock

ShimShmemManager

ShimShmemProcess

ShimShmemThread