shadow_shim_helper_rs::shim_shmem
Struct ProcessShmemProtected

#[repr(C)]
pub struct ProcessShmemProtected {
    pub host_id: HostId,
    pub pending_signals: sigset_t,
    /* private fields */
}

Fields
host_id: HostIdpending_signals: sigset_t
Implementations

impl ProcessShmemProtected

pub fn pending_standard_siginfo(&self, signal: Signal) -> Option<&siginfo_t>

pub fn set_pending_standard_siginfo(&mut self, signal: Signal, info: &siginfo_t)

pub unsafe fn clone_signal_actions(&mut self, src: &Self)

Safety

Only valid if pointers in src sigactions are valid in self’s address space (e.g. src is a parent that just forked this process).

pub unsafe fn signal_action(&self, signal: Signal) -> &sigaction

Safety

Function pointers in shd_kernel_sigaction::u are valid only from corresponding managed process, and may be libc::SIG_DFL or libc::SIG_IGN.

pub unsafe fn signal_action_mut(&mut self, signal: Signal) -> &mut sigaction

Safety

Function pointers in shd_kernel_sigaction::u are valid only from corresponding managed process, and may be libc::SIG_DFL or libc::SIG_IGN.

pub fn clear_pending_signals(&mut self)

This drops all pending signals. Intended primarily for use with exec.

pub fn take_pending_unblocked_signal( &mut self, thread: &ThreadShmemProtected, ) -> Option<(Signal, siginfo_t)>