shadow_shim_helper_rs::simulation_time

Struct SimulationTime

#[repr(C)]
pub struct SimulationTime(/* private fields */);

Implementations

impl SimulationTime

pub const MAX: SimulationTime

Maximum value. Currently equivalent to SIMTIME_MAX to avoid surprises when interoperating with C, but could use Duration::MAX when the C types go away.

pub const ZERO: SimulationTime

pub const SECOND: SimulationTime

pub const MILLISECOND: SimulationTime

pub const MICROSECOND: SimulationTime

pub const NANOSECOND: SimulationTime

pub fn from_c_simtime(val: CSimulationTime) -> Option<Self>

pub fn to_c_simtime(val: Option<Self>) -> CSimulationTime

pub const fn from_duration(val: Duration) -> Self

Convert a Duration to a SimulationTime. This function exists as a const alternative to SimulationTime::try_from(duration). May panic if the duration is too large.

pub fn is_zero(&self) -> bool

pub fn is_positive(&self) -> bool

pub fn as_secs(&self) -> u64

pub fn as_millis(&self) -> u64

pub fn as_micros(&self) -> u64

pub fn as_nanos(&self) -> u128

pub fn as_nanos_f64(&self) -> f64

pub fn checked_add(self, other: Self) -> Option<Self>

pub fn checked_sub(self, other: Self) -> Option<Self>

pub fn checked_mul(self, other: u64) -> Option<Self>

pub fn checked_div(self, other: u64) -> Option<Self>

pub fn checked_rem(self, other: Self) -> Option<Self>

pub fn saturating_add(self, other: Self) -> Self

pub fn saturating_sub(self, other: Self) -> Self

pub fn saturating_mul(self, other: u64) -> Self

pub fn try_from_secs(s: u64) -> Option<Self>

pub fn from_secs(s: u64) -> Self

pub fn try_from_millis(s: u64) -> Option<Self>

pub fn from_millis(s: u64) -> Self

pub fn try_from_micros(s: u64) -> Option<Self>

pub fn from_micros(s: u64) -> Self

pub fn try_from_nanos(s: u64) -> Option<Self>

pub fn from_nanos(s: u64) -> Self

pub fn subsec_millis(&self) -> u32

pub fn subsec_micros(&self) -> u32

pub fn subsec_nanos(&self) -> u32