Module shadow_shim_helper_rs::syscall_types

Structs

ForeignPtr

Represents a pointer to a virtual address in plugin memory. The pointer is not guaranteed to be valid or aligned.

ManagedPhysicalMemoryAddr

Represents a pointer to a physical address in plugin memory.

SysCallArgs

Type Aliases

UntypedForeignPtr

Used to indicate an untyped ForeignPtr in C code. We use the unit type as the generic rather than libc::c_void since the unit type is zero-sized and libc::c_void has a size of 1 byte, and this prevents us from accidentally accessing an “untyped” pointer.

Unions

SysCallReg

A register used for input/output in a syscall.